src/pyproject.toml
[project]
name = "jobfit-ai"
version = "0.0.0"
description = "Add your description here"
# HACK: Using Python 3.12 because `thinc` python package is not ready for 3.13
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "beanie>=1.29.0",
    "dependency-injector>=4.46.0",
    "fastapi[standard]>=0.115.11",
    "pdfplumber>=0.11.6",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "spacy>=3.8.5",
    "sqlmodel>=0.0.23",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=36.1.1",
    "pytest>=8.3.5",
    "reportlab>=4.4.0",
]
